+++ Sources/NyrisSDK/Requests/BaseAPIRequest.swift
@@ -14,8 +14,7 @@ public typealias OfferCompletion = (_ products:[Offer]?, _ error:Error?) -> Void
 /// Handle internet availability
 public class BaseService : NSObject {
     
-    public let endpointProvider:Endpoints
-    public let environmentMode:EnvironmentMode
+    public let endpointProvider:EndpointsProvider
     internal let client:NyrisClient = NyrisClient.instance
     internal let jsonTask:JSONDownloader
     public var currentTask:URLSessionTask?
@@ -46,15 +45,14 @@ public class BaseService : NSObject {
     /// instantiate a BaseAPIRequest in live mode
     ///
     /// - Parameter configuration: session configuration
-    public init(configuration:URLSessionConfiguration? = nil, environmentMode:EnvironmentMode = .live) {
+    public init(configuration:URLSessionConfiguration? = nil) {
         
-        self.environmentMode   = environmentMode
-        self.endpointProvider  = Endpoints(environmentMode: self.environmentMode)
+        self.endpointProvider  = client.endpointProvider
         
         if let configuration = configuration {
-            self.jsonTask = JSONDownloader(apiKey:NyrisClient.instance.clientID, configuration: configuration)
+            self.jsonTask = JSONDownloader(apiKey:client.clientID, configuration: configuration)
         } else {
-            self.jsonTask = JSONDownloader(apiKey:NyrisClient.instance.clientID)
+            self.jsonTask = JSONDownloader(apiKey:client.clientID)
         }
     }